calibration_with_blender/requirements.txt
numpy == 1.22.4
opencv-python == 4.6.0.66
seaborn == 0.11.2
PyYAML == 6.0
pycairo == 1.21.0
